pyproject.toml
[project]
name = "bloom"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "black>=25.9.0",
    "fastapi-users[sqlalchemy]>=15.0.1",
    "fastapi-users-db-sqlalchemy>=7.0.0",
    "mypy>=1.18.2",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.11.0",
    "pydantic[email]>=2.12.3",
    "pylint>=4.0.2",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
    "redis>=7.1.0",
    "bcrypt>=4.0.0",
    "httpx>=0.28.1",
    "anyio>=4.12.0",
    "ruff>=0.14.9",
    "pypdf>=4.0.0",
    "reportlab>=4.0.0",
]

[tool.black]
line-length = 88  # Максимальная длина строки
target-version = ['py313']  # Целевая версия Python
skip-string-normalization = true  # Не менять кавычки

[tool.pylint.main]
disable = "C0114,C0115,C0116"  # Отключаем предупреждения по документированию
max-line-length = 88  # Длина строки для Pylint

[tool.mypy]
files = "src"  # Папка с вашим кодом
disallow_untyped_calls = true  # Запрещает вызовы без типов

[tool.ruff]
# Включение и отключение определённых проверок
select = ["E", "F", "W"]  # Ошибки, предупреждения и другие проверки
ignore = []          # Игнорировать конкретные предупреждения